refactor separable solver
